also split out NodeType; consistently use is_unlinked()
