+++ src/tree/mod.rs
@@ -4,9 +4,10 @@
 pub mod document;
 pub mod namespace;
 pub mod node;
+pub mod nodetype;
 
 pub use tree::document::Document;
-pub use tree::namespace::Namespace;
-pub use tree::node::{Node, NodeType};
-
 pub(crate) use tree::document::DocumentRef;
+pub use tree::namespace::Namespace;
+pub use tree::node::Node;
+pub use tree::nodetype::NodeType;
